[Fix] 메뉴바 무한 랜더링 고치기 Close #366
